QQuickMenu: refactor all 'use native' logic into a single function
As it stood, the logic that resolved if a menu can, and
should be, native was duplicated into two functions;
useNativeMenu() and resolvedPopupType(). And the two was
somewhat inconsistent.

This patch will therefore refactor all the logic into
resolvedPopupType(), so that the popup type is resolved
in a single place. Existing auto tests
(such as tst_QQuickMenu::subMenuPopupType()) should verify
that this refactoring is safe.

Pick-to: 6.8
Change-Id: Ifc73d20c2d2e06646dc2849c8398660a35b17cb2
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>