QQuickDialog: reject on hideEvent, when popupType: Popup.Window
Dialogs should be either rejected or accepted, when they close.
When they close as a result of a hideEvent (usually when the user
presses the X button in the window decoration), it's expected that the
dialog emits the rejected signal.

QtQuickDialogs expects either accepted or rejected to be emitted when
the dialog closes. Because neither signal were emitted, QtQuickDialogs
would enter a state, where it falsly believe that the dialog is still
opened, when it was closed due to the hideEvent.

Fixes: QTBUG-129323
Change-Id: I03e0406c6c4ce0237ddfa625bb476b11013b7666
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit b09deeee514f2aae7251a55169c6c6af4777b0e9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit ff56ef92b3b5ea0c326e8ce1df230a6436c05082)