Log timings for individual GC steps
Add a logging category `qt.qml.gc.stepExecution` that is used to enable
the additional logs.

`GCStateMachine` was modified to carry some additional state related to
the registered timings over multiple run of a single step.

Call to the execution parts of a state were modified to collect some
measurements and log them if the relevant logging category is enabled.

Task-number: QTBUG-128357
Pick-to: 6.8
Change-Id: I6fc620c85534edc82546130f32bf2af49af6bcdd
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>