Add scale and transform notifiers
There are existing notifiers (QQuickItem::ItemChange and
QQuickItemChangeListener) for geometry (position and size), and rotation.
Scale has a public signal but no corresponding notifiers. Transform
is also affected by QQuickTransform instances attached to the item,
and when any of those change there is currently no mechanism available
to public or private API's to detect such changes. This change adds the
notifiers so the set is complete - Geometry, Rotation, Scale, and
Transform which catches all three and transform changes.

Task-number: QTBUG-10644
Change-Id: Ib57dbe23c600a493b4e08ce8edad078e10e28d09
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>