Fix BusyIndicator being hidden when running is changed quickly
If running is set to false and then true within a short period,
BusyIndicatorImpl's OpacityAnimator cancels the 1 => 0
animation (which was for running being set to false), setting opacity
to 0 and hence visible to false. This happens _after_ setRunning(true)
was called, because the properties were set synchronously but the
animation is asynchronous. To account for this situation, we only hide
ourselves if we're not running by storing and checking our running
state.

Fixes: QTBUG-85860
Pick-to: 6.7 6.5
Change-Id: I220dfb78f00028e4a12a92fc14082006e1844002
Reviewed-by: Doris Verria <doris.verria@qt.io>
(cherry picked from commit 15b385af9f8e24e8c66588b55a3f7828d70d2c33)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>