QQItem::nextPrevItemInTabFocusChain: Always respect wrap parameter
3f3787a6b4f54f2603598e49eeebb58532aa2afa added a wrap parameter to
QQItem::nextPrevItemInTabFocusChain that allows us to not wrap around
the focus chain for cases where eg: we need to tab navigate in and out
of the quick window. The problem was that we were wrongly wrapping
from the first back to the last in case of a BACKTAB, even if wrap was
set to false. This was to allow for cases where the root item has the
active focus and we want to backtab to the last item in the quick
window. This shouldn't be the case, however, because if wrap is
false, that means that we shouldn't wrap in ANY case. This was also
not allowing us to backtab to a parent window when the focus is on
the root item of the quick window.
To fix, if we are about to wrap around and the "wrap" parameter is
false, always return.

Change-Id: I29faa7c67bc229b9965b06acb4631b7c8e5b73d0
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 3b659a6a9e8592470eb07a46ca682b1459b7804c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>