Close Dialog when a DestructiveRole button is clicked
Assume that a button with DialogButtonBox.DestructiveRole is added to a
dialog. When that button is clicked, it only emitted the discarded()
signal. Now after emitting, we call QQuickDialog::close().

[ChangeLog][Controls][Important Behavior Changes] Dialog is now closed
when a button with the DestructiveRole is clicked. Currently the only
button with this role is Discard.

Fixes: QTBUG-67168
Change-Id: I3fbee383e10709630ba48cc420924ca71153e70f
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit fbc473d22a56b524e5dab769c573cbb3769aba50)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>