Doc: Add link to qmlformat page
To avoid duplicated content, add a link to the qmlformat page.
Having a link to a more detailed page for the tool is also in line with
how the other tools are listed on the Qt Quick Tools and Utilities page.

Fixes: QTBUG-127287
Pick-to: 6.7 6.5
Change-Id: Id0fc71818bdb41fb5763dad7f2b562e4c3a00a4d
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit e71fbeddf5f9e76d0191df4465a211aee89e951a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>