qmlls: add qmllint plugins to static builds
Remove the broken qt_import_plugins() call that was not doing anything,
and add a dependency to QmlCompiler to add qmllint plugins to static
qmlls builds.

The plugins are automatically collected and linked into static builds
using CMake-magic if the correspondings libraries are used for qmlls.

Therefore add QmlCompiler for the qmllint plugins as LIBRARIES to qmlls.

Technically, QmlCompiler is an indirect dependency as its actually
needed by QmlLSPrivate library.

Fixes: QTBUG-129603
Change-Id: Icfecebac8790d6e238531ae9bf4083e145d57f11
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>