QQuickFlickable: don't pause during drag when changing drag direction
Flickable/ListView/TableView would sometimes pause (or lag) for almost
a second if you dragged it in one direction, but then switched to
drag it in the opposite direction (without lifting your finger). This was
especially evident on iOS. It could be reproduced in, for example, the
Gallery app (the controls-list in the drawer), or any other app with
a Flickable.

The reason this happened, was because a "reversed" drag would cause
dx and dy to converge back to zero (compared to the starting point
of the drag), which would result in overThreshold becoming false
again. And since the deltas ended up below the threshold, the content
item stopped moving.

Rather than overThreshold becoming false when dragging back to the
starting point, this patch will change it so that it stays forever
true once dragging has started. It will also remove the
related dx/dy != 0 checks because of the same reason; Once in dragging
mode, the content item should always move, even if it was dragged back
to the starting point (dx/dy == 0).

Fixes: QTBUG-129599
Pick-to: 6.7 6.5 6.2
Change-Id: I86d4e6eba74bbb52884bac4782531ac90818323d
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
(cherry picked from commit dc5b8f5c4f126b7e25a169f24fbbfb570af09b39)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>