Curve renderer: ensure timely deallocation of temporary vertex data
During processing, a significant amount of intermediate ("uncooked")
vertex data is allocated for every fill and stroke node. After the
final, smaller vertex data is generated, the intermediate data was
intended to be deallocated right away. Since the QList API now
requires 2 calls to actually deallocate, that did not happen, and the
data would live on as long as the node (path) itself.

Pick-to: 6.8
Change-Id: I48efa5d0bc2b226c5eaac37dbba485b653604bfd
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>