QQmlSettings: Write parent properties from an overridden setting
`QQmlSettings` can save and load settings values based on the properties
defined on its usages in a QML project.

The logic that handles which properties are to be considered, currently
uses the MetaObject of a specific instance/type to filter out all
properties that are not defined by the instance/type itself.

It is possible to have a component derive from a `QQmlSettings`
component that defines certain properties.

With the current logic, all properties that are not directly defined on
the child component will not be considered when saving/loading settings
values, so that changes to them might be lost.

To fix the issue, ensure that instances of `QQmlSettings` consider all
properties that are not base properties of `QQmlSettings`, so that
inherited properties will be saved/loaded.

Add a test case to `tst_qqmlsettings` that show such a usage.

Fixes: QTBUG-129515
Change-Id: I3034de3cbe9a49969bae1ba90a567e4d0eda514a
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>