QmlCompiler: Untangle various list operations
We never need global types for those. The size of a list is a property
of that list, for list values we have a separate content variant, as
well as for list iterators. In order to determine whether something is a
list, we can query the type itself directly.

Task-number: QTBUG-124670
Change-Id: I7a81571b42cd0ec3662a4362c6caf98b22c9ebb1
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>