Memory manager: Ensure that internal class is set early
If the internal class is not set early enough, we will run into a crash
if the object to hold the internal class ends up on the markstack, and
the markstack gets drained. This can happen when the markstack runs out
of space (reaches the softlimit), and does early draining.
We force this condition in the test via a test-only function to
manipulate the soft-limit.

Pick-to: 6.8
Change-Id: I13598797811d697981ef7018c5aa87251331ab17
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>