qmlls: Fix crash on editing file
DocumentSymbols should work on the valid document.
Also, we might get invalid file, so we should check the validity of qml
dom file item before dereferencing it. This happens although we check
the validity of a file in the language server module before processing
the request. This is probably caused by a synchronization problem on
editing the file and creating the dom for that file. This should be
further investigated with QTBUG-121171.

Change-Id: I4a2582210deb8da78ee1400a9f2f3c84b3ecbbae
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>