Curve renderer: Avoid keeping unneeded QuadPath object
The pathData structure would keep a copy of the QuadPath used to
generate the stroke nodes. That copy was never used for anything, so
remove it.

This commit also cleans up the functions to generate stroke nodes, by
decoupling them from the internal PathData structure. This was earlier
done for the corresponding addFillNodes() function.

Pick-to: 6.8
Change-Id: I414b421dfc0cdf5c38e52a298476433f7643b5db
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>