Continue scrolling Flickable past start or end of nested Flickable
You can continue scrolling the main list with the mouse wheel even after
reaching the end of the nested list. When using a trackpad, behavior is
now the same.

Added tst_qquickflickable::nestedSameDirectionTrackpad.
nestedSameDirection.qml or the example from QTBUG-126514 is suitable
for manual testing.

Since we add a custom QPointingDevice to simulate a trackpad, use it in
nestedTrackpad() too, for better realism.

Fixes: QTBUG-124478
Pick-to: 6.5 6.7 6.8
Change-Id: I9b9c9a41afcfa5d950093a31682013ae9e917f1a
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>