QtQml: Remove base CU when clearing executable CU
This allows the base CU to be released while there may still be objects
around that reference the executable CU. Since the JS engine is gone,
the CU is not useful for much anymore anyway.

Since we don't want to check for the presence of a base CU in every
method, we replace the base CU with a static null value.

Fixes: QTBUG-129388
Change-Id: I84f8410fa8652203ac7c39f88a6d5aa05420ab11
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>