QtQml: Call lookups based on enums rather than via function pointers
While the C++ standard says that different functions need to have unique
addresses, some compilers have take substantial liberties with that
rule. This means we can't actually rely on the addresses of our
different lookup functions to differ and therefore we cannot use them as
discriminator.

Introduce an enumeration for all the different lookups and use that
instead.

Now we can also drop all the purely redirecting methods we've introduced
just to have different addresses.

Change-Id: Ifa68c27c0d2fef4084893a19227dab21bd948dfd
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>