QtQml: Protect lookupAttached() from identical COMDAT folding
MSVC thinks it can deviate from the standard and fold separate functions
into the same code block even if their addresses are taken. This makes
the addresses equal and breaks the lookup mechanism. We set up a dilemma
for the compiler by explicitly comparing the function addresses and
failing when they are equal. Now the new function has "different" code
that will however be optimized away.

Pick-to: 6.8 6.5 6.2
Fixes: QTBUG-128843
Change-Id: I42d9f764235b877c5c9e054502ea01a16369d27b
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>