QtQml: Protect value type lookupSetter() from identical COMDAT folding
MSVC thinks it can deviate from the standard and fold separate functions
into the same code block even if their addresses are taken. This makes
the addresses equal and breaks the lookup mechanism. We set up a dilemma
for the compiler by explicitly comparing the function addresses and
failing when they are equal. Now the new function has "different" code
that will however be optimized away.

Pick-to: 6.8 6.5 6.2
Task-number: QTBUG-128843
Change-Id: Id9512397fcbb238d880ee9966d8708fc576590c7
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>