Fix conflicting QGfxSourceProxy meta object with the one in qt5compat
There are two QGfxSourceProxy classes in the codebase, one in
qt5compat/src/imports/graphicaleffects5/private and the other in
qtdeclarative/src/effects.

In a static build, if a project links and initializes both qml
modules, it is undetermined which QGfxSourceProxy class' meta object
will be loaded.
This can lead to the multi effects object version being loaded when
initializing a component from the graphical effects module. You then
get a warning message saying "Missing QML.Element class info for
QGfxSourceProxy" and then a failure to load the QML module.

This happened for the sample qmake project attached to the bug report
this change is fixing.

Rename the multi effects QGfxSourceProxy class to QGfxSourceProxyME
(where ME stands for Multi Effects) to avoid the meta object conflict.

Pick-to: 6.8
Fixes: QTBUG-130080
Change-Id: I0001f36b19e36deeddd32897d23c8bfa52011e02
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>