qmllsutils: support call expressions in resolveExpressionType
Add a new IdentifierType called "NotAnIdentifier" which is used by
resolveExpressionType() for expressions, and use it when resolving the
type of callexpressions.

This fixes the recursive call of resolveExpressionType() when processing
`Qt.point().x` for example, as it now returns something when resolving
the left hand side `Qt.point()`.

This allows completion to suggest stuff after `Qt.point().`.

Pick-to: 6.8
Fixes: QTBUG-127776
Change-Id: I3cbee987c0677eb6d2b90fab44e982c49d0dd4fa
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>