CMake: Fix incorrect ctest argument for shared_qml_module test
The shared_qml_module cmake build test project does not have a
'shared_qml_module' inner test target, it only has a 'test_models'
one, so passing 'shared_qml_module' after the -V was incorrect, and
even if the test existed, it should have been passed to the -R
argument.

Unfortunately ctest versions prior to 3.30 happily ignore incorrect
arguments. This means that the -V argument was completely ignored
because it was treated as a single argument due to the double quotes.

Remove the incorrect test name, leaving just -V to run all found
tests, like we do for other build tests.

As a drive-by use the CMAKE_CTEST_COMMAND variable instead of plain
'ctest', like the other build tests.

This is needed to ensure CI passes when CMake is updated to 3.30.

Pick-to: 6.8
Task-number: QTQAINFRA-6484
Change-Id: Ic7d1c267985cf94d5806f3972d628ab1bf9352cf
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>