Defer polishAndSync animation advance to event loop if handling expose
If we are handling an expose event the system is expecting us to
produce a frame that it can present on screen to the user. Advancing
animations directly in polishAndSynct might result in changing properties
of the window in a way that invalidates the current frame, for example
the size of the window.

This might result in the discarding of the current frame before the user
ever sees it, either because the frame size is different than the window,
so it's determined to be no longer valid, or because the system never
gets a chance to present the frame because it starts another round of
expose events.

To give the system a chance to present the current frame we defer the
advance of the animations until the start of the next event loop pass,
which should still give plenty of time to compute the new render state
before the next expose event or update request.

We limit this approach for Apple operating systems for now, so that
we can pick it to 6.8. It will possibly be extended to other platforms
later on.

As the window can go away while we wait for the next event loop pass,
and even as part of advancing animations, we guard the window with
a QPointer.

Pick-to: 6.8
Fixes: QTBUG-129839
Change-Id: I32089599a8b6ba5c7177a92ff791f80e2812ffd8
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>