QmlCompiler: Make the ParentScope content variant more useful
It doesn't have to be JavaScript-specific. QML scopes also have parent
scopes that, in theory, we can retrieve and use. (we don't, for a
reason, but that reason doesn't have to be part of the name)

A scoped value is generally regarded as a "Property" of the scope, and
we can extend this meaning to JavaScript function or lexical scopes.
This makes the JavaScriptObject content variant obsolete.

Task-number: QTBUG-124670
Change-Id: I728f066461c000a0f2391f8604bd38853806adc0
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>