CMake: Qml plugins should inherit the STATIC-ness of their module
Otherwise in a shared build, a qml module marked as STATIC will
produce a static backing library and a dynamic qml plugin, which is
not something one usually expects, and can also lead to duplicate
symbol issues.

One such case was the QmlAssetDownloader qml module, which for a macOS
build ended up installing a libQt6QmlAssetDownloader.a together with
a libqmlassetdownloaderplugin.dylib , where the latter should have
also been a static library.

Pick-to: 6.8
Change-Id: If5e5c941cde49e7f2050b6a28b9db207dcddf771
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>