Don't propagate mouse event for menu configured inside popup
The patch e4a159632ec39542ec23567963191162f9dcdfef fixes propagating
mouse events by not allowing it to the underlying items beneath the
overlay.

It still can cause an issue when the menu is nested within the popup.
This is because the menu doesn't block the input, which allows the
overlay to pass on the events to the next child item.

This patch enables the menu to consume the mouse press event (as
similar to the drawer), further causing overlay not to pass on the
events to the other child items.

Fixes: QTBUG-130536
Change-Id: I1014c692a4a5aed22e096a79c06e27cb26fea70d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 9fd43c27d19e675a0f5efdda4ea53fa97f61f852)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit a5a54c083bbe8506a2f16ab8a411d3c8689c31d6)