TableView: ensure we position the syncView in the right direction
If a TableView has a syncView, a call to positionViewAtCell()
will instead be forwarded to the syncview, according to the
syncDirection. But as it stood, this was flipped, so that we
ended up forwarding a horizontal positioning call to a
vertically synced syncView, and vice versa.

This patch will correct this, so that the forwarded
positioning call matches the syncDirection.

Fixes: QTBUG-131776
Change-Id: Ibeaf5ce3d1d18c1ef0a0184c38336860fdee67be
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit bede04154deab7e8d7e992a76b272e37ecb41205)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 0011dad7f1f4f0945ae17876146619719b723e4d)