cmake: recreate .qmlls.ini on build folder change
Change the way that qmlls.ini files are generated in the source folder,
such that they can be automatically recreated on build folder change,
for example when using a source folder with multiple different build folders.

Introduce new files in the build folder: .qmlls.ini.build contains the
content of the .qmlls.ini, .qmlls.ini.timestamp is empty and is a marker
to indicate to CMake whether the .qmlls.ini.build content should be
written to the source folder .qmlls.ini.

When there is a timestamp in the build folder, then we know that the
.qmlls.ini file in the source folder is up-to-date and nothing needs to
be done.

If there is no timestamp (the project was not build before, or the
timestamp was deleted), use newly-added Qt6UpdateQmllsIni.cmake to
find the build folder used by the source-folder .qmlls.ini and to
invalidate (delete) all timestamp files it can find there, to signal
the previous build folder that its .qmlls.ini is out-dated.
Then, write the .qmlls.ini content to the source folder and create a
timestamp file in the current build folder.

If the previous build folder is built again, it will see the missing
timestamp files and will re-write the .qmlls.ini files in the source
folder.

Add a test in a separate commit because RunCMakeTest does not seem to
work yet.

Pick-to: 6.8
Task-number: QTBUG-133793
Change-Id: I66aa259e56d795860bb346d53d9841714048e478
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 538cd9d8ce5539d0b95d8e33cdd705a4cb388d53)
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>