qmlls: make QQmlJSImportVisitor process updates to files
Qmlls makes QQmlJSImportVisitor re-create a QQmlJSScope when a file gets
updated. It turns out that QQmlJSImportVisitor only works correctly when
used on an empty scopes, methods like rootScopeIsValid() don't work
correctly if the scope is not empty.

Therefore, reset the scopes before making QQmlJSImportVisitor run on
them. Make sure that the internalName and the moduleName are still set.

This fixes a bug where qmlls can't autocomplete an "in-memory" enum
because the import visitor does not set the new enum values in the root
element correctly on an non-empty scope.

Pick-to: 6.8
Fixes: QTCREATORBUG-32634
Change-Id: If6d620f350215074f87b53bb153363f2dec06145
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 72dc34f213d58f4fab6a699f5d5244c233e89878)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>