Doc: Avoid creating the qqmlintegration.h and qqml.h proxy pages
QDoc allows positioning some amount of documentation under certain pages
by use of the "\relates" command.

The "\relates" command takes an argument that represents a name and
tries to find where the documentation should be positioned in the
output, based on that name.

When the provided name doesn't refer to anything that QDoc is aware of
in the current project, a "proxy page" will be created to hold the
documentation.

The "qqmlintegration.h" header is documented as an header and its
generated page was intended to be used to collect the documentation for
a series of elements that the header exports.

Nonetheless, while the header page itself was created while, the
elements that were intended to be collected where insted positioned
under a similarly named proxy page.

The issues stemmed from the "\headerfile" command using the
"<qqmlintegration.h>" name to refer to the header and the relevant
relates command using the "qqmlintegration.h" name to refer to the same,
with the two names differing in the usage of the surrounding angle
brackets.

To avoid creating a proxy page, and instead correctly collecting the
relevant documentation under the header documentation page, the name
usage was synchronized by changing the relevant "\relates" command to
refer to "<qqmlintegration.h>".

A similar issue existes for the "qqml.h" headerfile and a similar
solution was applied to solve the issue.

Fixes: QTBUG-136037
Change-Id: I3bf1aa97648c0a94acb7a0de5f5fbbaeb6a81c7b
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit aefd01697767bda646c597ff774d87e899436ee1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 88f340b07181ec98e7859672e21daf76defecc68)