Map DragHandler delta to the scene
Previously, DragHandler didn't account for scene transforms when
calculating drag threshold. In the case when one axis is disabled,
and the scene is rotated 90 degrees, the remaining axis should be
activated by movements that are nominally the same as the disabled
direction, because of rotation.

Fixes: QTBUG-136354
Pick-to: 6.8
Change-Id: I6669949908b0d026850411c10056685648a109d9
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 93f7f3afdb461265ccd65b7bd79761817b52dc55)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>