QQmlProperty: Fix read of value type lists
When we introduced value type lists, we didn't adjust the QQmlProperty
code which assumed that any list property could be represented as
QQmlListProperty.

This was of course no longer the case, and attempting to use it with
such a property would result in UB, and most likely in crashes.

Fix this by detecting what kind of list we are dealing with.

Pick-to: 6.8 6.5
Fixes: QTBUG-136735
Change-Id: I1e4531da3e5f26994809f02d630820541849691c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit e96ad30e430a95f8cb40379f08f0166d5ac3c855)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>