QSGThreadedRenderLoop: fix race on window pointer
When the thread is not running yet, we indeed have to set the window
pointer. And we can do that without mutex/atomic because starting the
thread is a synchronization point.

When the thread is already running, however, writing into 'window'
races with reads of that variable in the thread. But it's useful to do
that, after the thread handled a WM_Obscure event which set the window
pointer to nullptr (testcase: tst_qquickwindow destroyShowWithoutHide).
We need to set it again if the window is exposed again: use an event
to do this in a thread-safe way.

WARNING: ThreadSanitizer: data race (pid=427588)
  Read of size 8 at 0x723400006048 by thread T11:
    #0 QSGRenderThread::syncAndRender() qsgthreadedrenderloop.cpp:802
  Previous write of size 8 at 0x723400006048 by main thread:
    #0 QSGThreadedRenderLoop::handleExposure(QQuickWindow*) qsgthreadedrenderloop.cpp:1294

Pick-to: 6.8 6.5
Change-Id: I145bb499628fc0c95c4362a32aecdc7c3d688677
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit f4dddb5b6e47a08ffe9d000aad586b781b4fe82f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>