tst_qquickmenu::contextMenuKeyboard: Fix flakiness
The test case assumes that QQuickMenu will handle the key press
events giving the focus items the proper focus reason (Tab/Backtab),
instead of the internal ListView handling them.
However, in most styles, the 'interactive' property of the Menu's
ListView is bound to a condition depending on the 'height' of the Menu,
and if evaluated to true, would cause the ListView to be interactive and
handle key events directly. Make sure that the test assumptions are
valid by explicitly setting and checking for this in the test.

Fixes: QTBUG-133858
Pick-to: 6.8
Change-Id: Iad458ba96167404a40605ddfbb5b8d2d83d94590
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 52f95ea50fad029f38fb2bfa64b4c58dfacbdc10)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>