ComboBox: fix crash when setting model
In certain use cases, setting a ListModel that's created by a Loader
can cause a heap-use-after-free. Account for this by storing a
QPointer to the model if it's QObject*, and checking the validity
of that before casting.

I wasn't able to reproduce this in an auto test.

Rename some local variables while we're at it, to improve the
readability of the code.

Fixes: QTBUG-136672
Pick-to: 6.8
Change-Id: I84bbf21b37fb5675c931600ef8ea47365cceb09a
Reviewed-by: Christopher Adams <chris.adams@qinetic.com.au>
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit be034ebd5c982b1cda7fe4254795193ef4972525)
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>